#159 - refactors to be more canjs-like
