+++ .travis.yml
@@ -4,8 +4,5 @@ before_install:
   - ./bin/firefox-allow-popups.sh
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
-before_script:
-  - npm install -g bower
-  - bower install
 script: npm test
 sudo: false
