--- documentjs.json
@@ -1,18 +1,0 @@
-{
-	"sites": {
-		"docs": {
-			"parent": "FuncUnit",
-			"pageConfig": {"page":"docs"},
-			"glob": {
-				"pattern": "{browser/**/*.{js,md},README.md}"
-			}
-		},
-		"guides":{
-			"glob": {
-				"pattern": "docs/guides/**/*.{js,md}"
-			},
-			"parent": "Guides",
-			"pageConfig": {"page":"guides"}
-		}
-	}
-}
