+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "funcunit",
-  "main": "dist/cjs/funcunit.js",
   "version": "3.1.0-pre.2",
   "author": {
     "name": "Bitovi",
@@ -10,31 +9,7 @@
   "repository": {
     "type": "git",
     "url": "git@github.com:bitovi/funcunit.git",
-    "github": "https://github.com/bitovi/funcunit",
-    "publish": "git push origin && git push origin --tags",
-    "release:prerelease": "npm version prerelease && npm publish",
-    "release:patch": "npm version patch && npm publish",
-    "release:minor": "npm version minor && npm publish",
-    "release:major": "npm version major && npm publish"
-  },
-  "devDependencies": {
-    "bower": "^1.3.8",
-    "can-zone": "^0.4.4",
-    "chai": "^3.5.0",
-    "documentjs": "^0.3.0-pre.4",
-    "grunt": "0.4.x",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-concat": "0.3.0",
-    "steal": "^0.15.1",
-    "steal-jasmine": "0.0.2",
-    "steal-mocha": "0.0.3",
-    "steal-qunit": "^0.1.1",
-    "steal-tools": "^0.15.4",
-    "testee": "^0.2.5"
-  },
-  "scripts": {
-    "test": "grunt testee",
-    "prepublish": "./node_modules/.bin/grunt build"
+    "github": "https://github.com/bitovi/funcunit"
   },
   "licenses": [
     {
@@ -42,9 +17,50 @@
       "url": "http://opensource.org/licenses/mit-license.php"
     }
   ],
+  "scripts": {
+    "preversion": "npm test && npm run build",
+    "version": "git commit -am \"Update dist for release\" && git checkout -b release && git add -f dist/",
+    "postversion": "git push --tags && git checkout master && git branch -D release && git push",
+    "testee": "npm run testee:funcunit && npm run testee:none && npm run testee:qunit && npm run testee:mocha && npm run testee:jasmine && npm run testee:jasmine2",
+    "testee:funcunit": "testee test.html --browsers firefox",
+    "testee:none": "testee test/adapters/none.html --browsers firefox",
+    "testee:qunit": "testee test/adapters/qunit.html --browsers firefox",
+    "testee:mocha": "testee test/adapters/mocha.html --browsers firefox",
+    "testee:jasmine": "testee test/adapters/jasmine.html --browsers firefox",
+    "testee:jasmine2": "testee test/adapters/jasmine2.html --browsers firefox",
+    "test": "npm run jshint && npm run testee",
+    "jshint": "jshint *.js --config",
+    "release:patch": "npm version patch && npm publish",
+    "release:minor": "npm version minor && npm publish",
+    "release:major": "npm version major && npm publish",
+    "build": "node build.js",
+    "document": "bit-docs",
+    "develop": "done-serve --static --develop --port 8080",
+    "release:pre": "npm version prerelease && npm publish"
+  },
+  "main": "dist/cjs/funcunit.js",
   "dependencies": {
     "jquery": "1.11.0",
+    "jquerypp": "^2.0.2",
     "syn": "^0.2.0"
+  },
+  "devDependencies": {
+    "bit-docs": "0.0.6",
+    "can-define": "^0.8.0",
+    "can-zone": "^0.4.4",
+    "chai": "^3.5.0",
+    "documentjs": "^0.3.0-pre.4",
+    "done-serve": "^0.2.5",
+    "donejs-cli": "^0.9.5",
+    "generator-donejs": "^0.9.0",
+    "jasmine": "^2.5.2",
+    "jshint": "^2.9.1",
+    "steal": "^0.16.35",
+    "steal-jasmine": "0.0.3",
+    "steal-mocha": "0.0.3",
+    "steal-qunit": "^0.1.1",
+    "steal-tools": "^0.16.0",
+    "testee": "^0.2.5"
   },
   "system": {
     "main": "funcunit",
@@ -52,12 +68,7 @@
       "chai": "chai/chai"
     },
     "npmIgnore": [
-      "bower",
       "documentjs",
-      "grunt",
-      "grunt-cli",
-      "grunt-contrib-concat",
-      "grunt-contrib-connect",
       "steal",
       "steal-tools",
       "testee"
