+++ test/drag/drag.html
@@ -33,8 +33,6 @@
 			<div id="drop"></div>
 		</div>
 		<div class="status"></div>
-        <script src="../../lib/jquery/dist/jquery.js"></script>
-        <script src="../jquerypp.js"></script>
 		<script src="../../node_modules/steal/steal.js" main="funcunit/test/drag/drag"></script>
 	</body>
 </html>
