+++ test/scroll.html
@@ -3,16 +3,6 @@
 	<head>
 		<meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
 		<title>Untitled Document</title>
-   <script src="../lib/jquery/dist/jquery.js" type='text/javascript'></script>
-   <script type='text/javascript'>
-       $(function(){
-				$('#innerdiv').click(function()
-            {
-              $('#scrolldiv').scrollTop(100);
-              $('#scrolldiv').scrollLeft(100);
-            })
-       })
-		</script>
 	</head>
 	<body>
 		<h1>Scroll Test</h1>
@@ -20,5 +10,15 @@
          <div id='scrolldiv' style='width:100px;height:200px;overflow-y:scroll;' tabindex='0'>
          <div id='innerdiv' style='width:500px; height:1000px;'><a id="scrollme" href='javascript://'>click to scroll</a></div> 
       </div>
+
+			<script src="../node_modules/steal/steal.js">
+	        var $ = require("funcunit/browser/jquery");
+	        $(function(){
+						$('#innerdiv').click(function(){
+		          $('#scrolldiv').scrollTop(100);
+		          $('#scrolldiv').scrollLeft(100);
+		        });
+	        });
+			</script>
 		</body>
 </html>
